Fusion behavior and entity again
